chore: move ChatViewModel to ui package
